Added a new WRF MCS tracking workflow.
1. Added a new MCS tracking run script for WRF: run_mcs_tbpf_wrf.py. The run script performs end-to-end MCS tracking on raw WRF output data using Tb and rain rate.

2. Modified preprocess_wrf_tb_rainrate.py to a function and integrated into run_mcs_tbpf_wrf.py.

3. Added a flag ‘run_preprocess’ in config_wrf4km_mcs_tbpf_example.yml to control preprocessing the WRF data to produce Tb and rain rate from raw WRF output files.

4. Moved grow_cells and get_neighborhood functions from label_and_grow_cold_clouds.py to ftfunctions.py. These two general functions can be used in other codes.

5. Corrected a bug introduced in the last commit in run_mcs_tbpf.py.
