+++ config/config_wrf4km_mcs_tbpf_example.yml
@@ -3,6 +3,7 @@
 # Tracking uses collocated Tb + Precipitation
 
 # Processing steps:
+run_preprocess: False
 run_idfeature : True
 run_tracksingle : True
 run_gettracks : True
@@ -50,7 +51,7 @@ landmask_varname: 'LANDMASK'
 # Input dataset structure
 pixel_radius:  4.0  # [km] Spatial resolution of the input data
 datatimeresolution: 1.0  # [hour] Temporal resolution of the input data
-# Variable names in the input dat
+# Variable names in the input data
 tb_varname:  'tb'
 pcp_varname: 'rainrate'
 clouddatasource: 'model'
