+++ pyflextrkr/preprocess_wrf_tb_rainrate.py
@@ -1,22 +1,74 @@
-"""
-Preprocess WRF output to make infrared brightness temperature and rain rate
-for tracking deep convective clouds.
-
-To run this code:
-python preprocess_wrf_tb_rainrate.py config.yml
-"""
 import numpy as np
 import time
 import os, sys, glob
 import logging
-from netCDF4 import Dataset
 import xarray as xr
 import pandas as pd
-from wrf import (getvar, ALL_TIMES)
+from netCDF4 import Dataset
+from wrf import getvar, ALL_TIMES
 from itertools import repeat
 from multiprocessing import Pool
-from pyflextrkr.ft_utilities import load_config
+from pyflextrkr.ftfunctions import olr_to_tb
 
+def preprocess_wrf_tb_rainrate(config):
+    """
+    Preprocess WRF output file to get Tb and rain rate.
+
+    Args:
+        config: dictionary
+            Dictionary containing config parameters
+    
+    Returns:
+        None.
+    """
+
+    # Set the logging message level
+    logger = logging.getLogger(__name__)
+    
+    # Get inputs from config
+    run_parallel = config['run_parallel']
+    n_workers = config['nprocesses']
+    indir = config['wrfout_path']
+    outdir = config['clouddata_path']
+    inbasename = config['wrfout_basename']
+    outbasename = config['databasename']
+
+    # Create output directory
+    os.makedirs(outdir, exist_ok=True)
+
+    # Find all WRF files
+    filelist = sorted(glob.glob(f'{indir}/{inbasename}*'))
+    nfiles = len(filelist)
+    logger.info(f'Number of WRF files: {nfiles}')
+
+    # Create a list with a pair of WRF filenames that are adjacent in time
+    filepairlist = []
+    for ii in range(0, nfiles-1):
+        ipair = [filelist[ii], filelist[ii+1]]
+        filepairlist.append(ipair)
+    nfilepairs = len(filepairlist)
+    
+    if run_parallel == 0:
+        # Serial version
+        for ifile in range(0, nfilepairs):
+            status = calc_rainrate_tb(filepairlist[ifile], outdir, inbasename, outbasename)
+    elif run_parallel == 1:
+        # Parallel version
+        # Use starmap to unpack the iterables as arguments
+        # For arguments that are the same for each iterable, use itertools "repeat" to duplicate those
+        # Example follows this stackoverflow post
+        # https://stackoverflow.com/questions/5442910/python-multiprocessing-pool-map-for-multiple-arguments
+        # Refer to Pool.starmap:
+        # https://docs.python.org/dev/library/multiprocessing.html#multiprocessing.pool.Pool.starmap
+        pool = Pool(n_workers)
+        pool.starmap(calc_rainrate_tb, zip(filepairlist, repeat(outdir), repeat(inbasename), repeat(outbasename)) )
+        pool.close()
+    else:
+        sys.exit('Valid parallelization flag not set.')
+        
+    return
+
+    
 def calc_rainrate_tb(filepairnames, outdir, inbasename, outbasename):
     """
     Calculates rain rates from a pair of WRF output files and write to netCDF
@@ -36,6 +88,7 @@ def calc_rainrate_tb(filepairnames, outdir, inbasename, outbasename):
         Returns status = 1 if success.
     """
 
+    logger = logging.getLogger(__name__)
     status = 0
     
     # Filenames with full path
@@ -111,16 +164,8 @@ def calc_rainrate_tb(filepairnames, outdir, inbasename, outbasename):
     ncfile_t1.close()
     ncfile_t2.close()
     
-    # Calculate brightness temperature
-    # (1984) as given in Yang and Slingo (2001)
-    # Tf = tb(a+b*Tb) where a = 1.228 and b = -1.106e-3 K^-1
-    # OLR = sigma*Tf^4 
-    # where sigma = Stefan-Boltzmann constant = 5.67x10^-8 W m^-2 K^-4
-    a = 1.228
-    b = -1.106e-3
-    sigma = 5.67e-8 # W m^-2 K^-4
-    tf = (OLR/sigma)**0.25
-    tb = (-a + np.sqrt(a**2 + 4*b*tf))/(2*b)
+    # Convert OLR to IR brightness temperature
+    tb = olr_to_tb(OLR)
 
     # Write single time frame to netCDF output
     for tt in range(0, ntimes-1):
@@ -139,7 +184,7 @@ def calc_rainrate_tb(filepairnames, outdir, inbasename, outbasename):
         }
         gattr_dict = {
             'Title': 'WRF calculated rainrate and brightness temperature',
-            'contact': 'Zhe Feng: zhe.feng@pnnl.gov',
+            'Contact': 'Zhe Feng: zhe.feng@pnnl.gov',
             'Institution': 'Pacific Northwest National Laboratory',
             'created on': time.ctime(time.time()),
             'Original_File1': filein_t1,
@@ -152,6 +197,7 @@ def calc_rainrate_tb(filepairnames, outdir, inbasename, outbasename):
         # Specify attributes
         dsout['time'].attrs['long_name'] = 'Epoch time (seconds since 1970-01-01 00:00:00)'
         dsout['time'].attrs['units'] = 'seconds since 1970-01-01 00:00:00'
+        dsout['time'].attrs['_FillValue'] = np.NaN
         dsout['Times'].attrs['long_name'] = 'WRF-based time'
         dsout['lon2d'].attrs['long_name'] = 'Longitude'
         dsout['lon2d'].attrs['units'] = 'degrees_east'
@@ -165,7 +211,7 @@ def calc_rainrate_tb(filepairnames, outdir, inbasename, outbasename):
         # Write to netcdf file
         encoding_dict = {
             # 'base_time': {'zlib':True, 'dtype':'int64'},
-            'time':{'zlib':True, 'dtype':'int64'},
+            'time':{'zlib':True, 'dtype':'float'},
             'Times':{'zlib':True},
             'lon2d':{'zlib':True, 'dtype':'float32'},
             'lat2d':{'zlib':True, 'dtype':'float32'},
@@ -178,54 +224,5 @@ def calc_rainrate_tb(filepairnames, outdir, inbasename, outbasename):
         return (status)
 
 
-if __name__ == "__main__":
-
-    # Set the logging message level
-    logging.basicConfig(format="%(asctime)s - %(name)s - %(levelname)s - %(message)s", level=logging.INFO)
-    logger = logging.getLogger(__name__)
-
-    # Load configuration file
-    config_file = sys.argv[1]
-    config = load_config(config_file)
-    # Get inputs from config
-    run_parallel = config['run_parallel']
-    n_workers = config['nprocesses']
-    indir = config['wrfout_path']
-    outdir = config['clouddata_path']
-    inbasename = config['wrfout_basename']
-    outbasename = config['databasename']
-
-    # Create output directory
-    os.makedirs(outdir, exist_ok=True)
-
-    # Find all WRF files
-    filelist = sorted(glob.glob(f'{indir}/{inbasename}*'))
-    nfiles = len(filelist)
-    logger.info(f'Number of WRF files: {nfiles}')
-
-    # Create a list with a pair of WRF filenames that are adjacent in time
-    filepairlist = []
-    for ii in range(0, nfiles-1):
-        ipair = [filelist[ii], filelist[ii+1]]
-        filepairlist.append(ipair)
-    nfilepairs = len(filepairlist)
-    
-    if run_parallel == 0:
-        # Serial version
-        for ifile in range(0, nfilepairs):
-            status = calc_rainrate_tb(filepairlist[ifile], outdir, inbasename, outbasename)
-    elif run_parallel == 1:
-        # Parallel version
-        # Use starmap to unpack the iterables as arguments
-        # For arguments that are the same for each iterable, use itertools "repeat" to duplicate those
-        # Example follows this stackoverflow post
-        # https://stackoverflow.com/questions/5442910/python-multiprocessing-pool-map-for-multiple-arguments
-        # Refer to Pool.starmap:
-        # https://docs.python.org/dev/library/multiprocessing.html#multiprocessing.pool.Pool.starmap
-        pool = Pool(n_workers)
-        pool.starmap(calc_rainrate_tb, zip(filepairlist, repeat(outdir), repeat(inbasename), repeat(outbasename)) )
-        pool.close()
-    else:
-        sys.exit('Valid parallelization flag not set.')
